+++ concordia/admin/__init__.py
@@ -46,77 +46,19 @@ from .actions import (
     unpublish_action,
     unpublish_item_action,
 )
-from .filters import AcceptedFilter, RejectedFilter, SubmittedFilter
+from .filters import (
+    AcceptedFilter,
+    AssetProjectListFilter2,
+    ItemProjectListFilter2,
+    RejectedFilter,
+    SubmittedFilter,
+    TranscriptionProjectListFilter,
+)
 from .forms import (
     AdminItemImportForm,
     BleachedDescriptionAdminForm,
     SimpleContentBlockAdminForm,
 )
-
-
-class CampaignProjectListFilter(admin.SimpleListFilter):
-    """
-    This filter is an example of how to combine two different Filters to work together.
-    """
-
-    # Title displayed on the list filter URL
-    title = "ProjectRedux"
-    # Model field name:
-    parameter_name = "project"
-    # Custom attributes
-    related_filter_parameter = "project__campaign__id__exact"
-
-    def lookups(self, request, model_admin):
-        list_of_questions = []
-        queryset = Project.objects.order_by("campaign_id")
-        if self.related_filter_parameter in request.GET:
-            queryset = queryset.filter(
-                campaign_id=request.GET[self.related_filter_parameter]
-            )
-        for project in queryset:
-            list_of_questions.append((str(project.id), project.title))
-        return sorted(list_of_questions, key=lambda tp: tp[1])
-
-    def queryset(self, request, queryset):
-        # Compare the requested value to decide how to filter the queryset.
-        if self.value():
-            return queryset.filter(project_id=self.value())
-        return queryset
-
-
-class ItemProjectListFilter2(CampaignProjectListFilter):
-    parameter_name = "project__in"
-
-
-class ProjectItemListFilter(admin.SimpleListFilter):
-    #  modification of filter above but for asset through item
-    # Title displayed on the list filter URL
-    title = "ProjectRedux"
-    # Model field name:
-    parameter_name = "project"
-    # Custom attributes
-    related_filter_parameter = "item__project__campaign__id__exact"
-
-    def lookups(self, request, model_admin):
-        list_of_questions = []
-        queryset = Project.objects.order_by("campaign_id")
-        if self.related_filter_parameter in request.GET:
-            queryset = queryset.filter(
-                campaign_id=request.GET[self.related_filter_parameter]
-            )
-        for project in queryset:
-            list_of_questions.append((str(project.id), project.title))
-        return sorted(list_of_questions, key=lambda tp: tp[1])
-
-    def queryset(self, request, queryset):
-        # Compare the requested value to decide how to filter the queryset.
-        if self.value():
-            return queryset.filter(item__project_id=self.value())
-        return queryset
-
-
-class AssetProjectListFilter2(ProjectItemListFilter):
-    parameter_name = "item__project__in"
 
 
 class ProjectListFilter(MultipleChoiceListFilter):
@@ -557,7 +499,8 @@ class TranscriptionAdmin(admin.ModelAdmin):
         AcceptedFilter,
         RejectedFilter,
         "asset__item__project__campaign",
-        "asset__item__project",
+        # "asset__item__project",
+        TranscriptionProjectListFilter,
     )
 
     search_fields = ["text", "user__username", "user__email"]
